+++ src/lib/ericchase/Utility/Console.ts
@@ -1,65 +1,64 @@
-namespace Console {
-  export let newline_count = 0;
-  export let marks = new Set<{ updated: boolean }>();
-}
+import { UpdateMarkerManager } from './UpdateMarker.js';
 
-function updateMarks() {
-  for (const mark of Console.marks) {
-    Console.marks.delete(mark);
-    mark.updated = true;
-  }
-}
+const marker_manager = new UpdateMarkerManager<{ newline_count: number }>({ newline_count: 0 });
 
-export function GetConsoleMark() {
-  const mark = { updated: false };
-  Console.marks.add(mark);
-  return mark;
+export function GetConsoleMarker() {
+  return marker_manager.getNewMarker();
 }
 
 export function ConsoleError(...items: any[]) {
   // biome-ignore lint: this let's us search for undesired console[error]s
   console['error'](...items);
-  Console.newline_count = 0;
-  updateMarks();
+  marker_manager.extra.newline_count = 0;
+  marker_manager.updateMarkers();
 }
 
 export function ConsoleErrorWithDate(...items: any[]) {
   // biome-ignore lint: this let's us search for undesired console[error]s
   console['error'](`[${new Date().toLocaleTimeString()}]`, ...items);
-  Console.newline_count = 0;
-  updateMarks();
+  marker_manager.extra.newline_count = 0;
+  marker_manager.updateMarkers();
 }
 
 export function ConsoleLog(...items: any[]) {
   // biome-ignore lint: this let's us search for undesired console[log]s
   console['log'](...items);
-  Console.newline_count = 0;
-  updateMarks();
+  marker_manager.extra.newline_count = 0;
+  marker_manager.updateMarkers();
 }
 
 export function ConsoleLogWithDate(...items: any[]) {
   // biome-ignore lint: this let's us search for undesired console[log]s
   console['log'](`[${new Date().toLocaleTimeString()}]`, ...items);
-  Console.newline_count = 0;
-  updateMarks();
+  marker_manager.extra.newline_count = 0;
+  marker_manager.updateMarkers();
 }
 
 export function ConsoleNewline(ensure_count = 1) {
-  for (let i = Console.newline_count; i < ensure_count; i++) {
+  for (let i = marker_manager.extra.newline_count; i < ensure_count; i++) {
     // biome-ignore lint: this let's us search for undesired console[log]s
     console['log']();
-    Console.newline_count++;
+    marker_manager.extra.newline_count++;
   }
-  updateMarks();
+  marker_manager.updateMarkers();
 }
 
 export function ConsoleLogToLines(items: Iterable<any>) {
-  for (const item of items) {
-    ConsoleLog(item);
+  if (typeof items === 'string') {
+    ConsoleLog(items);
+  } else {
+    for (const item of items) {
+      ConsoleLog(item);
+    }
   }
 }
+
 export function ConsoleErrorToLines(items: Iterable<any>) {
-  for (const item of items) {
-    ConsoleError(item);
+  if (typeof items === 'string') {
+    ConsoleError(items);
+  } else {
+    for (const item of items) {
+      ConsoleError(item);
+    }
   }
 }