+++ src/lib/ericchase/Utility/HelpMessage.ts
@@ -1,0 +1,11 @@
+import { ConsoleError } from './Console.js';
+import { ImmediateDebounce } from './Debounce.js';
+
+export class HelpMessage {
+  print: () => Promise<void>;
+  constructor(public message: string) {
+    this.print = ImmediateDebounce(() => {
+      ConsoleError(this.message);
+    }, 500);
+  }
+}