+++ src/lib/ericchase/Utility/TextCodec.ts
@@ -1,7 +1,7 @@
 const textDecoder = new TextDecoder();
 const textEncoder = new TextEncoder();
 
-export function DecodeText(buffer: Uint8Array) {
+export function DecodeBytes(buffer: Uint8Array) {
   return textDecoder.decode(buffer);
 }
 export function EncodeText(text: string) {