+++ tools/lib/preprocessors/FilePreprocessor_IIFEWrapper.ts
@@ -1,5 +1,5 @@
 import type { Path, PathGroup } from '../../../src/lib/ericchase/Platform/Node/Path.js';
-import { DecodeText, EncodeText } from '../../../src/lib/ericchase/Utility/TextCodec.js';
+import { DecodeBytes, EncodeText } from '../../../src/lib/ericchase/Utility/TextCodec.js';
 import type { FilePreprocessor } from './FilePreprocessor.js';
 
 export class IIFEWrapperPreprocessor implements FilePreprocessor {
@@ -16,6 +16,6 @@ export class IIFEWrapperPreprocessor implements FilePreprocessor {
     return false;
   }
   preprocess(bytes: Uint8Array) {
-    return { bytes: EncodeText(`(() => {\n${DecodeText(bytes)}})();`) };
+    return { bytes: EncodeText(`(() => {\n${DecodeBytes(bytes)}})();`) };
   }
 }